Bring in libuuid and libshare

+++ ZFSin/zfs/module/unicode/u8_textprep.c
@@ -45,6 +45,7 @@
 #include <sys/ddi.h>
 #include <sys/sunddi.h>
 #else
+#include <sys/w32_types.h>
 #include <sys/u8_textprep.h>
 #include <strings.h>
 #endif	/* _KERNEL */
@@ -336,7 +337,7 @@ const uint8_t u8_valid_max_2nd_byte[0x100] = {
  * specific to UTF-8 and Unicode.
  */
 int
-u8_validate(char *u8str, size_t n, char **list, int flag, int *errnum)
+u8_validate(char *u8str, uint32_t n, char **list, int flag, int *errnum)
 {
 	uchar_t *ib;
 	uchar_t *ibtail;
@@ -345,7 +346,7 @@ u8_validate(char *u8str, size_t n, char **list, int flag, int *errnum)
 	uchar_t *s2;
 	uchar_t f;
 	int sz;
-	size_t i;
+	uint32_t i;
 	int ret_val;
 	boolean_t second;
 	boolean_t no_need_to_validate_entire;
@@ -460,18 +461,18 @@ u8_validate(char *u8str, size_t n, char **list, int flag, int *errnum)
  * The return value is the byte length of the returned character excluding
  * the terminating null byte.
  */
-static size_t
+static uint32_t
 do_case_conv(int uv, uchar_t *u8s, uchar_t *s, int sz, boolean_t is_it_toupper)
 {
-	size_t i;
+	uint32_t i;
 	uint16_t b1 = 0;
 	uint16_t b2 = 0;
 	uint16_t b3 = 0;
 	uint16_t b3_tbl;
 	uint16_t b3_base;
 	uint16_t b4 = 0;
-	size_t start_id;
-	size_t end_id;
+	uint32_t start_id;
+	uint32_t end_id;
 
 	/*
 	 * At this point, the only possible values for sz are 2, 3, and 4.
@@ -507,23 +508,23 @@ do_case_conv(int uv, uchar_t *u8s, uchar_t *s, int sz, boolean_t is_it_toupper)
 	 */
 	b1 = u8_common_b1_tbl[uv][b1];
 	if (b1 == U8_TBL_ELEMENT_NOT_DEF)
-		return ((size_t)sz);
+		return ((uint32_t)sz);
 
 	b2 = u8_case_common_b2_tbl[uv][b1][b2];
 	if (b2 == U8_TBL_ELEMENT_NOT_DEF)
-		return ((size_t)sz);
+		return ((uint32_t)sz);
 
 	if (is_it_toupper) {
 		b3_tbl = u8_toupper_b3_tbl[uv][b2][b3].tbl_id;
 		if (b3_tbl == U8_TBL_ELEMENT_NOT_DEF)
-			return ((size_t)sz);
+			return ((uint32_t)sz);
 
 		start_id = u8_toupper_b4_tbl[uv][b3_tbl][b4];
 		end_id = u8_toupper_b4_tbl[uv][b3_tbl][b4 + 1];
 
 		/* Either there is no match or an error at the table. */
 		if (start_id >= end_id || (end_id - start_id) > U8_MB_CUR_MAX)
-			return ((size_t)sz);
+			return ((uint32_t)sz);
 
 		b3_base = u8_toupper_b3_tbl[uv][b2][b3].base;
 
@@ -532,13 +533,13 @@ do_case_conv(int uv, uchar_t *u8s, uchar_t *s, int sz, boolean_t is_it_toupper)
 	} else {
 		b3_tbl = u8_tolower_b3_tbl[uv][b2][b3].tbl_id;
 		if (b3_tbl == U8_TBL_ELEMENT_NOT_DEF)
-			return ((size_t)sz);
+			return ((uint32_t)sz);
 
 		start_id = u8_tolower_b4_tbl[uv][b3_tbl][b4];
 		end_id = u8_tolower_b4_tbl[uv][b3_tbl][b4 + 1];
 
 		if (start_id >= end_id || (end_id - start_id) > U8_MB_CUR_MAX)
-			return ((size_t)sz);
+			return ((uint32_t)sz);
 
 		b3_base = u8_tolower_b3_tbl[uv][b2][b3].base;
 
@@ -550,7 +551,7 @@ do_case_conv(int uv, uchar_t *u8s, uchar_t *s, int sz, boolean_t is_it_toupper)
 	 * If i is still zero, that means there is no corresponding character.
 	 */
 	if (i == 0)
-		return ((size_t)sz);
+		return ((uint32_t)sz);
 
 	u8s[i] = '\0';
 
@@ -567,15 +568,15 @@ do_case_conv(int uv, uchar_t *u8s, uchar_t *s, int sz, boolean_t is_it_toupper)
  * faster processing time.
  */
 static int
-do_case_compare(size_t uv, uchar_t *s1, uchar_t *s2, size_t n1,
-	size_t n2, boolean_t is_it_toupper, int *errnum)
+do_case_compare(uint32_t uv, uchar_t *s1, uchar_t *s2, uint32_t n1,
+	uint32_t n2, boolean_t is_it_toupper, int *errnum)
 {
 	int f;
 	int sz1;
 	int sz2;
-	size_t j;
-	size_t i1;
-	size_t i2;
+	uint32_t j;
+	uint32_t i1;
+	uint32_t i2;
 	uchar_t u8s1[U8_MB_CUR_MAX + 1];
 	uchar_t u8s2[U8_MB_CUR_MAX + 1];
 
@@ -692,7 +693,7 @@ do_case_compare(size_t uv, uchar_t *s1, uchar_t *s2, size_t n1,
  * a Starter.
  */
 static uchar_t
-combining_class(size_t uv, uchar_t *s, size_t sz)
+combining_class(uint32_t uv, uchar_t *s, uint32_t sz)
 {
 	uint16_t b1 = 0;
 	uint16_t b2 = 0;
@@ -744,8 +745,8 @@ combining_class(size_t uv, uchar_t *s, size_t sz)
  * the null byte. It will also return a state that will tell if there was
  * a Hangul character decomposed which then will be used by the caller.
  */
-static size_t
-do_decomp(size_t uv, uchar_t *u8s, uchar_t *s, int sz,
+static uint32_t
+do_decomp(uint32_t uv, uchar_t *u8s, uchar_t *s, int sz,
 	boolean_t canonical_decomposition, u8_normalization_states_t *state)
 {
 	uint16_t b1 = 0;
@@ -754,9 +755,9 @@ do_decomp(size_t uv, uchar_t *u8s, uchar_t *s, int sz,
 	uint16_t b3_tbl;
 	uint16_t b3_base;
 	uint16_t b4 = 0;
-	size_t start_id;
-	size_t end_id;
-	size_t i;
+	uint32_t start_id;
+	uint32_t end_id;
+	uint32_t i;
 	uint32_t u1;
 
 	if (sz == 2) {
@@ -851,15 +852,15 @@ do_decomp(size_t uv, uchar_t *u8s, uchar_t *s, int sz,
 	/* Try to find matching decomposition mapping byte sequence. */
 	b1 = u8_common_b1_tbl[uv][b1];
 	if (b1 == U8_TBL_ELEMENT_NOT_DEF)
-		return ((size_t)sz);
+		return ((uint32_t)sz);
 
 	b2 = u8_decomp_b2_tbl[uv][b1][b2];
 	if (b2 == U8_TBL_ELEMENT_NOT_DEF)
-		return ((size_t)sz);
+		return ((uint32_t)sz);
 
 	b3_tbl = u8_decomp_b3_tbl[uv][b2][b3].tbl_id;
 	if (b3_tbl == U8_TBL_ELEMENT_NOT_DEF)
-		return ((size_t)sz);
+		return ((uint32_t)sz);
 
 	/*
 	 * If b3_tbl is bigger than or equal to U8_16BIT_TABLE_INDICATOR
@@ -877,7 +878,7 @@ do_decomp(size_t uv, uchar_t *u8s, uchar_t *s, int sz,
 
 	/* This also means there wasn't any matching decomposition. */
 	if (start_id >= end_id)
-		return ((size_t)sz);
+		return ((uint32_t)sz);
 
 	/*
 	 * The final table for decomposition mappings has three types of
@@ -933,7 +934,7 @@ do_decomp(size_t uv, uchar_t *u8s, uchar_t *s, int sz,
 	if (canonical_decomposition) {
 		/* Is the mapping only for compatibility decomposition? */
 		if (b1 < U8_DECOMP_BOTH)
-			return ((size_t)sz);
+			return ((uint32_t)sz);
 
 		start_id++;
 
@@ -968,7 +969,7 @@ do_decomp(size_t uv, uchar_t *u8s, uchar_t *s, int sz,
  * to the composition mappings as explained in the do_composition().
  */
 static uchar_t *
-find_composition_start(size_t uv, uchar_t *s, size_t sz)
+find_composition_start(uint32_t uv, uchar_t *s, uint32_t sz)
 {
 	uint16_t b1 = 0;
 	uint16_t b2 = 0;
@@ -976,8 +977,8 @@ find_composition_start(size_t uv, uchar_t *s, size_t sz)
 	uint16_t b3_tbl;
 	uint16_t b3_base;
 	uint16_t b4 = 0;
-	size_t start_id;
-	size_t end_id;
+	uint32_t start_id;
+	uint32_t end_id;
 
 	if (sz == 1) {
 		b4 = s[0];
@@ -1035,10 +1036,10 @@ find_composition_start(size_t uv, uchar_t *s, size_t sz)
  * characters in this sequence and return whether it is blocked or not.
  */
 static boolean_t
-blocked(uchar_t *comb_class, size_t last)
+blocked(uchar_t *comb_class, uint32_t last)
 {
 	uchar_t my_comb_class;
-	size_t i;
+	uint32_t i;
 
 	my_comb_class = comb_class[last];
 	for (i = 1; i < last; i++)
@@ -1056,25 +1057,25 @@ blocked(uchar_t *comb_class, size_t last)
  *
  * The input argument 's' cannot contain more than 32 characters.
  */
-static size_t
-do_composition(size_t uv, uchar_t *s, uchar_t *comb_class, uchar_t *start,
-	uchar_t *disp, size_t last, uchar_t **os, uchar_t *oslast)
+static uint32_t
+do_composition(uint32_t uv, uchar_t *s, uchar_t *comb_class, uchar_t *start,
+	uchar_t *disp, uint32_t last, uchar_t **os, uchar_t *oslast)
 {
 	uchar_t t[U8_STREAM_SAFE_TEXT_MAX + 1];
 	uchar_t tc[U8_MB_CUR_MAX];
 	uint8_t saved_marks[U8_MAX_CHARS_A_SEQ];
-	size_t saved_marks_count;
+	uint32_t saved_marks_count;
 	uchar_t *p;
 	uchar_t *saved_p;
 	uchar_t *q;
-	size_t i;
-	size_t saved_i;
-	size_t j;
-	size_t k;
-	size_t l;
-	size_t C;
-	size_t saved_l;
-	size_t size;
+	uint32_t i;
+	uint32_t saved_i;
+	uint32_t j;
+	uint32_t k;
+	uint32_t l;
+	uint32_t C;
+	uint32_t saved_l;
+	uint32_t size;
 	uint32_t u1;
 	uint32_t u2;
 	boolean_t match_not_found = B_TRUE;
@@ -1378,8 +1379,8 @@ SAFE_RETURN:
  * including 0. The return value does not include the terminating
  * null byte.
  */
-static size_t
-collect_a_seq(size_t uv, uchar_t *u8s, uchar_t **source, uchar_t *slast,
+static uint32_t
+collect_a_seq(uint32_t uv, uchar_t *u8s, uchar_t **source, uchar_t *slast,
 	boolean_t is_it_toupper,
 	boolean_t is_it_tolower,
 	boolean_t canonical_decomposition,
@@ -1390,18 +1391,18 @@ collect_a_seq(size_t uv, uchar_t *u8s, uchar_t **source, uchar_t *slast,
 	uchar_t *s;
 	int sz;
 	int saved_sz;
-	size_t i;
-	size_t j;
-	size_t k;
-	size_t l;
+	uint32_t i;
+	uint32_t j;
+	uint32_t k;
+	uint32_t l;
 	uchar_t comb_class[U8_MAX_CHARS_A_SEQ];
 	uchar_t disp[U8_MAX_CHARS_A_SEQ];
 	uchar_t start[U8_MAX_CHARS_A_SEQ];
 	uchar_t u8t[U8_MB_CUR_MAX];
 	uchar_t uts[U8_STREAM_SAFE_TEXT_MAX + 1];
 	uchar_t tc;
-	size_t last;
-	size_t saved_last;
+	uint32_t last;
+	uint32_t saved_last;
 	uint32_t u1;
 
 	/*
@@ -1714,7 +1715,7 @@ TURN_STREAM_SAFE:
 
 	*source = s;
 
-	return ((size_t)sz);
+	return ((uint32_t)sz);
 }
 
 /*
@@ -1727,12 +1728,12 @@ TURN_STREAM_SAFE:
  * The meanings on the return values are the same as the usual strcmp().
  */
 static int
-do_norm_compare(size_t uv, uchar_t *s1, uchar_t *s2, size_t n1, size_t n2,
+do_norm_compare(uint32_t uv, uchar_t *s1, uchar_t *s2, uint32_t n1, uint32_t n2,
 	int flag, int *errnum)
 {
 	int result;
-	size_t sz1;
-	size_t sz2;
+	uint32_t sz1;
+	uint32_t sz2;
 	uchar_t u8s1[U8_STREAM_SAFE_TEXT_MAX + 1];
 	uchar_t u8s2[U8_STREAM_SAFE_TEXT_MAX + 1];
 	uchar_t *s1last;
@@ -1843,12 +1844,12 @@ do_norm_compare(size_t uv, uchar_t *s1, uchar_t *s2, size_t n1, size_t n2,
  * can be requested and checked against.
  */
 int
-u8_strcmp(const char *s1, const char *s2, size_t n, int flag, size_t uv,
+u8_strcmp(const char *s1, const char *s2, uint32_t n, int flag, uint32_t uv,
 		int *errnum)
 {
 	int f;
-	size_t n1;
-	size_t n2;
+	uint32_t n1;
+	uint32_t n2;
 
 	*errnum = 0;
 
@@ -1912,9 +1913,9 @@ u8_strcmp(const char *s1, const char *s2, size_t n, int flag, size_t uv,
 	    flag, errnum));
 }
 
-size_t
-u8_textprep_str(char *inarray, size_t *inlen, char *outarray, size_t *outlen,
-	int flag, size_t unicode_version, int *errnum)
+uint32_t
+u8_textprep_str(char *inarray, uint32_t *inlen, char *outarray, uint32_t *outlen,
+	int flag, uint32_t unicode_version, int *errnum)
 {
 	int f;
 	int sz;
@@ -1929,28 +1930,28 @@ u8_textprep_str(char *inarray, size_t *inlen, char *outarray, size_t *outlen,
 	boolean_t canonical_decomposition;
 	boolean_t compatibility_decomposition;
 	boolean_t canonical_composition;
-	size_t ret_val;
-	size_t i;
-	size_t j;
+	uint32_t ret_val;
+	uint32_t i;
+	uint32_t j;
 	uchar_t u8s[U8_STREAM_SAFE_TEXT_MAX + 1];
 	u8_normalization_states_t state;
 
 	if (unicode_version > U8_UNICODE_LATEST) {
 		*errnum = ERANGE;
-		return ((size_t)-1);
+		return ((uint32_t)-1);
 	}
 
 	f = flag & (U8_TEXTPREP_TOUPPER | U8_TEXTPREP_TOLOWER);
 	if (f == (U8_TEXTPREP_TOUPPER | U8_TEXTPREP_TOLOWER)) {
 		*errnum = EBADF;
-		return ((size_t)-1);
+		return ((uint32_t)-1);
 	}
 
 	f = flag & (U8_CANON_DECOMP | U8_COMPAT_DECOMP | U8_CANON_COMP);
 	if (f && f != U8_TEXTPREP_NFD && f != U8_TEXTPREP_NFC &&
 	    f != U8_TEXTPREP_NFKD && f != U8_TEXTPREP_NFKC) {
 		*errnum = EBADF;
-		return ((size_t)-1);
+		return ((uint32_t)-1);
 	}
 
 	if (inarray == NULL || *inlen == 0)
@@ -1958,7 +1959,7 @@ u8_textprep_str(char *inarray, size_t *inlen, char *outarray, size_t *outlen,
 
 	if (outarray == NULL) {
 		*errnum = E2BIG;
-		return ((size_t)-1);
+		return ((uint32_t)-1);
 	}
 
 	ib = (uchar_t *)inarray;
@@ -1990,7 +1991,7 @@ u8_textprep_str(char *inarray, size_t *inlen, char *outarray, size_t *outlen,
 			if (sz < 0) {
 				if (do_not_ignore_invalid) {
 					*errnum = EILSEQ;
-					ret_val = (size_t)-1;
+					ret_val = (uint32_t)-1;
 					break;
 				}
 
@@ -2001,7 +2002,7 @@ u8_textprep_str(char *inarray, size_t *inlen, char *outarray, size_t *outlen,
 			if (sz == 1) {
 				if (ob >= obtail) {
 					*errnum = E2BIG;
-					ret_val = (size_t)-1;
+					ret_val = (uint32_t)-1;
 					break;
 				}
 
@@ -2016,13 +2017,13 @@ u8_textprep_str(char *inarray, size_t *inlen, char *outarray, size_t *outlen,
 			} else if ((ib + sz) > ibtail) {
 				if (do_not_ignore_invalid) {
 					*errnum = EINVAL;
-					ret_val = (size_t)-1;
+					ret_val = (uint32_t)-1;
 					break;
 				}
 
 				if ((obtail - ob) < (ibtail - ib)) {
 					*errnum = E2BIG;
-					ret_val = (size_t)-1;
+					ret_val = (uint32_t)-1;
 					break;
 				}
 
@@ -2041,7 +2042,7 @@ u8_textprep_str(char *inarray, size_t *inlen, char *outarray, size_t *outlen,
 
 					if ((obtail - ob) < i) {
 						*errnum = E2BIG;
-						ret_val = (size_t)-1;
+						ret_val = (uint32_t)-1;
 						break;
 					}
 
@@ -2052,7 +2053,7 @@ u8_textprep_str(char *inarray, size_t *inlen, char *outarray, size_t *outlen,
 				} else {
 					if ((obtail - ob) < sz) {
 						*errnum = E2BIG;
-						ret_val = (size_t)-1;
+						ret_val = (uint32_t)-1;
 						break;
 					}
 
@@ -2085,7 +2086,7 @@ u8_textprep_str(char *inarray, size_t *inlen, char *outarray, size_t *outlen,
 			    ((ib + 1) < ibtail && U8_ISASCII(*(ib + 1))))) {
 				if (ob >= obtail) {
 					*errnum = E2BIG;
-					ret_val = (size_t)-1;
+					ret_val = (uint32_t)-1;
 					break;
 				}
 
@@ -2111,13 +2112,13 @@ u8_textprep_str(char *inarray, size_t *inlen, char *outarray, size_t *outlen,
 				    errnum, &state);
 
 				if (*errnum && do_not_ignore_invalid) {
-					ret_val = (size_t)-1;
+					ret_val = (uint32_t)-1;
 					break;
 				}
 
 				if ((obtail - ob) < j) {
 					*errnum = E2BIG;
-					ret_val = (size_t)-1;
+					ret_val = (uint32_t)-1;
 					break;
 				}
